+++ lib/src/row_data.dart
@@ -1,0 +1,7 @@
+class RowData<ROW> {
+  RowData({required this.row, required this.index, required this.hovered});
+
+  final ROW row;
+  final int index;
+  final bool hovered;
+}